feat(cloud-dedicated): v1 API primer - part of #4863, #4864 Still untested against CST: - v1 auth methods - v1 /write   - Telegraf config   - Precision   - API sample code
