+++ layouts/shortcodes/api/cloud/v2-prefer.html
@@ -1,1 +1,1 @@
-Use the v2 API [`/api/v2/write` endpoint](/influxdb/cloud-dedicated/write-data/v2-write-api/) for all new workloads.
+For new workloads, use the [InfluxDB v2 API `/api/v2/write` and `/api/v2/query` endpoints](/influxdb/cloud-dedicated/primers/api/v2/).
