+++ shared/text/api/cloud-dedicated/basic-auth.sh
@@ -1,13 +1,11 @@
 #######################################
-# Use an InfluxDB 1.x compatible username and password
-# to query the InfluxDB 1.x compatibility API
+# Use Basic Authentication with a Database Token
+# to query the InfluxDB v1 API
 #######################################
-# Use authentication query parameters:
-#   ?p=DATABASE_TOKEN
+# Use the --user option with `--user username:DATABASE_TOKEN` syntax
 #######################################
 
 curl --get "http://localhost:8086/query" \
-  --data-urlencode "u=[USERNAME]" \
-  --data-urlencode "p=DATABASE_TOKEN" \
+  --user "":"DATABASE_TOKEN" \
   --data-urlencode "db=DATABASE_NAME" \
   --data-urlencode "q=SELECT * FROM MEASUREMENT"
